feat(store/premiumize): improve types

+++ store/premiumize/account.go
@@ -1,10 +1,10 @@
 package premiumize
 
 type GetAccountInfoData struct {
-	CustomerId   string `json:"customer_id"`
-	PremiumUntil int    `json:"premium_until"`
-	LimitUsed    int    `json:"limit_used"`
-	SpaceUsed    int    `json:"space_used"`
+	CustomerId   string  `json:"customer_id"`
+	PremiumUntil int     `json:"premium_until"`
+	LimitUsed    float32 `json:"limit_used"`
+	SpaceUsed    int     `json:"space_used"`
 }
 
 type getAccountInfoData struct {
